grpc-client/requirements.txt
grpcio-tools==1.62.2
grpcio==1.62.2
protobuf>=4.21.6
flask
gunicorn>=22.0.0